+++ src/main/java/com/cells/cells/hyperdensity/fluid/ItemFluidHyperDensityCellBase.java
@@ -21,6 +21,7 @@ import appeng.api.storage.ICellInventoryHandler;
 import appeng.api.storage.channels.IFluidStorageChannel;
 import appeng.api.storage.data.IAEFluidStack;
 import appeng.fluids.helper.FluidCellConfig;
+import appeng.util.ReadableNumberConverter;
 
 import com.cells.cells.common.AbstractTieredCellItem;
 import com.cells.cells.common.INBTSizeProvider;
@@ -89,8 +90,9 @@ public abstract class ItemFluidHyperDensityCellBase extends AbstractTieredCellIt
             }
 
             if (CellUpgradeHelper.hasEqualDistributionUpgrade(getUpgradesInventory(stack))) {
-                long per_type = this.getBytesPerType(stack);
-                tooltip.add("§b" + I18n.format("tooltip.cells.upgrade.per_type", per_type));
+                long perType = this.getBytesPerType(stack);
+                String perTypeShort = ReadableNumberConverter.INSTANCE.toWideReadableForm(perType);
+                tooltip.add("§b" + I18n.format("tooltip.cells.upgrade.per_type", perType, perTypeShort));
             }
 
             // Add NBT size information (if enabled in config)